Career Copilot/requirements.txt
# Core dependencies
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
pdfplumber>=0.10.0
spacy>=3.7.0
nltk>=3.8.0
numpy>=1.24.0

# RAG dependencies
faiss-cpu>=1.7.4

# API dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# UI dependencies
streamlit>=1.28.0
requests>=2.31.0

# LLM dependencies (optional - install based on preference)
openai>=1.3.0  # For OpenAI GPT
google-genai>=1.0.0  # For Google Gemini

# Additional utilities
python-dotenv>=1.0.0
